Switch decoder bit timing limits to floating point values in units of µs

+++ include/rtl_433.h
@@ -62,10 +62,10 @@ struct protocol_state {
 
     unsigned int modulation;
 
-    /* pwm limits */
-    int short_limit;
-    int long_limit;
-    int reset_limit;
+    /* pwm limits (provided by driver in µs and converted to samples) */
+    float short_limit;
+    float long_limit;
+    float reset_limit;
     char *name;
     unsigned long demod_arg;
 };
